requirements.txt
# Core scientific packages
numpy
pandas
matplotlib

# ANSYS packages
ansys-mapdl-core
ansys-mapdl-reader

# Configuration
pyyaml

# Development tools
pytest
black
pylint
